requirements.txt
asgiref==3.3.1
Django==3.1.7
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
numpy==1.19.5
pandas==1.1.5
pkg-resources==0.0.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
